Reorganize several files
