+++ src/Notifications.API/Database/AppDbContext.cs
@@ -3,29 +3,12 @@ using Notifications.Models;
 
 namespace Notifications.Database;
 
-internal static class AppDbContextExtensions
+internal static class DatabaseExtensions
 {
     public static void ConfigureDatabase(this WebApplicationBuilder builder)
     {
         builder.Services.AddDbContext<AppDbContext>(options =>
-        {
-            options.UseSqlite(builder.Configuration.GetConnectionString("DefaultConnection"));
-
-            if (builder.Environment.IsDevelopment())
-            {
-                options.UseAsyncSeeding((context, _, token) => TestData.SeedDataAsync((AppDbContext)context, token));
-            }
-        });
-    }
-
-    public static async Task BuildDatabaseInDevAsync(this WebApplication app)
-    {
-        if (!app.Environment.IsDevelopment()) return;
-
-        using var scope = app.Services.CreateScope();
-        var context = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-        await context.Database.EnsureDeletedAsync();
-        await context.Database.EnsureCreatedAsync();
+            options.UseSqlite(builder.Configuration.GetConnectionString("DefaultConnection")));
     }
 }
