+++ src/Notifications.API/Platform/Endpoints.cs
@@ -1,7 +1,6 @@
 using Notifications.Database;
-using Notifications.Settings;
 
-namespace Notifications.Endpoints;
+namespace Notifications.Platform;
 
 internal static class Endpoints
 {
