+++ tests/Notifications.API.Tests/GetNotificationsTests.cs
@@ -13,7 +13,7 @@ public class GetNotificationsTests
     {
         var options = new DbContextOptionsBuilder<AppDbContext>()
             .UseInMemoryDatabase(databaseName: "RepositoryTest")
-            .UseAsyncSeeding((context, _, token) => TestData.SeedDataAsync((AppDbContext)context, token))
+            .UseAsyncSeeding((context, _, _) => TestData.SeedDataAsync((AppDbContext)context))
             .Options;
 
         _dbContext = new AppDbContext(options);
